Added more protection and fixed merge

+++ Trabalho/list/deleteList.php
@@ -3,6 +3,10 @@
   include(__DIR__ . '/../database/connection.php');
  try{
 
+    if (!preg_match ("/^[a-zA-Z\s]+$/", $_POST['listName'])){
+       header("Location: deleteListPage.php?specialChars");
+       die();
+    }
     $stmt = $dbh->prepare('SELECT idCategory FROM Category WHERE category = ?');
     $stmt->execute(array($_POST['listName']));
     $row = $stmt->fetch();
@@ -11,10 +15,13 @@
     $stmt = $dbh->prepare('DELETE FROM ELEMENT WHERE idCategory = ?');
     $stmt->execute(array($id));
 
-    $stmt = $dbh->prepare('DELETE FROM category WHERE category = ?');
-    $stmt->execute(array($_POST['listName']));
-    header("Location: ../index.php");
-
+    if ($stmt->fetch() != null){
+      $stmt = $dbh->prepare('DELETE FROM category WHERE category = ?');
+      $stmt->execute(array($_POST['listName']));
+      header("Location: ../index.php");
+   }
+   else
+     header("Location: deleteListPage.php?categoryNotFound");
    }
   catch (Exception $e) {
     echo 'Caught exception: ',  $e->getMessage(), "\n";
